Downloads/RAG-Getrepaired/requirements.txt
sentence-transformers
chromadb
groq
streamlit
python-dotenv
rank-bm25
pypdf
